Move sendOrder scheduling coverage to new test
