+++ webtransport/handlers/sendorder.py
@@ -1,34 +1,52 @@
-import json
+from typing import Optional, Tuple
+from urllib.parse import urlsplit, parse_qsl
 
+return_stream_id = 0;
+summary : bytes = [];
 
 def session_established(session):
-    session.dict_for_handlers['streams'] = {}
-
+    # When a WebTransport session is established, a bidirectional stream is
+    # created by the server, which is used to echo back stream data from the
+    # client.
+    path: Optional[bytes] = None
+    for key, value in session.request_headers:
+        if key == b':path':
+            path = value
+    assert path is not None
+    qs = dict(parse_qsl(urlsplit(path).query))
+    token = qs[b'token']
+    if token is None:
+        raise Exception('token is missing, path = {}'.format(path))
+    session.dict_for_handlers['token'] = token
+    global summary;
+    # need an initial value to replace
+    session.stash.put(key=token, value=summary)
 
 def stream_data_received(session,
                          stream_id: int,
                          data: bytes,
                          stream_ended: bool):
-    streams = session.dict_for_handlers['streams']
-    stream = streams.setdefault(stream_id, {
-        'marker': None,
-        'bytes_received': 0,
-    })
+    # we want to record the order that data arrives, and feed that ordering back to
+    # the sender.  Instead of echoing all the data, we'll send back
+    # just the first byte of each message.   This requires the sender to send buffers
+    # filled with only a single byte value.
+    # The test can then examine the stream of data received by the server to
+    # determine if orderings are correct.
+    # note that the size in bytes received here can vary wildly
 
-    if data:
-        if stream['marker'] is None:
-            stream['marker'] = data[0]
-        stream['bytes_received'] += len(data)
+    # Send back the data on the control stream
+    global summary
+    summary += data[0:1]
+    token = session.dict_for_handlers['token']
+    old_data = session.stash.take(key=token) or {}
+    session.stash.put(key=token, value=summary)
 
-    if not stream_ended:
-        return
+def stream_reset(session, stream_id: int, error_code: int) -> None:
+    global summary;
+    token = session.dict_for_handlers['token']
+    session.stash.put(key=token, value=summary)
+    summary = []
 
-    counts = {
-        str(item['marker']): item['bytes_received']
-        for item in streams.values()
-        if item['marker'] is not None
-    }
-    session.send_stream_data(
-        stream_id,
-        json.dumps(counts).encode('utf-8'),
-        end_stream=True)
+# do something different to include datagrams...
+def datagram_received(session, data: bytes):
+    session.send_datagram(data)
